requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
gradio>=3.50.0
pillow>=9.0.0
numpy>=1.21.0